split guards tests
